Add files via upload
This project is quite the same as Party Planer but bit more "In-depth" or advanced, or what ever you want to call it. same concept, choose a title for the project, set a cost and fee per participant, and create the event list.
Add the mount of Participants you want, can be unlimited, set the first and last name, street info and what country they should be from, and they will appear in the list.
List can also change and delete each participant.